go: refactor hash field expiration commands and fix return types
- Change HSetEx return type from bool to int64 to match Redis protocol
- Move command argument builders from options package to internal package
- Fix conditional option handling to only include when explicitly set
- Update documentation formatting for better readability
- Update all tests and examples to use new int64 return type
- Consolidate hash field expiration logic in internal/command_helpers.go
