+++ go/base_client.go
@@ -1843,7 +1843,8 @@ func (client *baseClient) HRandFieldWithCountWithValues(ctx context.Context, key
 //
 // Return value:
 //
-//	True if the operation was successful, false otherwise.
+//   - 1 if all fields were set successfully.
+//   - 0 if no fields were set due to conditional restrictions.
 //
 // [valkey.io]: https://valkey.io/commands/hsetex/
 func (client *baseClient) HSetEx(
@@ -1851,22 +1852,18 @@ func (client *baseClient) HSetEx(
 	key string,
 	fieldsAndValues map[string]string,
 	opts options.HSetExOptions,
-) (bool, error) {
-	args, err := options.BuildHSetExArgs(key, fieldsAndValues, opts)
+) (int64, error) {
+	args, err := internal.BuildHSetExArgs(key, fieldsAndValues, opts)
 	if err != nil {
-		return false, err
+		return 0, err
 	}
 
 	result, err := client.executeCommand(ctx, C.HSetEx, args)
 	if err != nil {
-		return false, err
+		return 0, err
 	}
 
-	intResult, err := handleIntResponse(result)
-	if err != nil {
-		return false, err
-	}
-	return intResult == 1, nil
+	return handleIntResponse(result)
 }
 
 // Gets the values of one or more fields of a given hash key and optionally sets their expiration time or time-to-live
@@ -1887,9 +1884,9 @@ func (client *baseClient) HSetEx(
 //
 // Return value:
 //
-//	An array of [models.Result[string]] values associated with the given fields, in the same order as they are requested.
-//	For every field that does not exist in the hash, a [models.CreateNilStringResult()] is returned.
-//	If key does not exist, returns an empty string array.
+//   - An array of [models.Result[string]] values associated with the given fields, in the same order as they are requested.
+//   - For every field that does not exist in the hash, a [models.CreateNilStringResult()] is returned.
+//   - If key does not exist, returns an empty string array.
 //
 // [valkey.io]: https://valkey.io/commands/hgetex/
 func (client *baseClient) HGetEx(
@@ -1898,7 +1895,7 @@ func (client *baseClient) HGetEx(
 	fields []string,
 	opts options.HGetExOptions,
 ) ([]models.Result[string], error) {
-	args, err := options.BuildHGetExArgs(key, fields, opts)
+	args, err := internal.BuildHGetExArgs(key, fields, opts)
 	if err != nil {
 		return nil, err
 	}
@@ -1939,11 +1936,11 @@ func (client *baseClient) HGetEx(
 //
 // Return value:
 //
-//	An array of integers indicating the result for each field:
-//		-2: Field does not exist in the hash, or key does not exist.
-//		0: The specified condition was not met.
-//		1: The expiration time was applied.
-//		2: When called with 0 seconds.
+//   - An array of integers indicating the result for each field:
+//   - -2: Field does not exist in the hash, or key does not exist.
+//   - 0: The specified condition was not met.
+//   - 1: The expiration time was applied.
+//   - 2: When called with 0 seconds.
 //
 // [valkey.io]: https://valkey.io/commands/hexpire/
 func (client *baseClient) HExpire(
@@ -1953,7 +1950,7 @@ func (client *baseClient) HExpire(
 	fields []string,
 	opts options.HExpireOptions,
 ) ([]int64, error) {
-	args, err := options.BuildHExpireArgs(key, expireTime, fields, opts, false)
+	args, err := internal.BuildHExpireArgs(key, expireTime, fields, opts, false)
 	if err != nil {
 		return nil, err
 	}
@@ -1986,11 +1983,11 @@ func (client *baseClient) HExpire(
 //
 // Return value:
 //
-//	An array of integers indicating the result for each field:
-//		-2: Field does not exist in the hash, or hash is empty.
-//		0: The specified condition was not met.
-//		1: The expiration time was applied.
-//		2: When called with 0 seconds or past Unix time.
+//   - An array of integers indicating the result for each field:
+//   - -2: Field does not exist in the hash, or hash is empty.
+//   - 0: The specified condition was not met.
+//   - 1: The expiration time was applied.
+//   - 2: When called with 0 seconds or past Unix time.
 //
 // [valkey.io]: https://valkey.io/commands/hexpireat/
 func (client *baseClient) HExpireAt(
@@ -2000,7 +1997,7 @@ func (client *baseClient) HExpireAt(
 	fields []string,
 	opts options.HExpireOptions,
 ) ([]int64, error) {
-	args, err := options.BuildHExpireArgs(key, expireTime, fields, opts, false)
+	args, err := internal.BuildHExpireArgs(key, expireTime, fields, opts, false)
 	if err != nil {
 		return nil, err
 	}
@@ -2032,11 +2029,11 @@ func (client *baseClient) HExpireAt(
 //
 // Return value:
 //
-//	An array of integers indicating the result for each field:
-//		-2: Field does not exist in the hash, or hash is empty.
-//		0: The specified condition was not met.
-//		1: The expiration time was applied.
-//		2: When called with 0 milliseconds.
+//   - An array of integers indicating the result for each field:
+//   - -2: Field does not exist in the hash, or hash is empty.
+//   - 0: The specified condition was not met.
+//   - 1: The expiration time was applied.
+//   - 2: When called with 0 milliseconds.
 //
 // [valkey.io]: https://valkey.io/commands/hpexpire/
 func (client *baseClient) HPExpire(
@@ -2046,7 +2043,7 @@ func (client *baseClient) HPExpire(
 	fields []string,
 	opts options.HExpireOptions,
 ) ([]int64, error) {
-	args, err := options.BuildHExpireArgs(key, expireTime, fields, opts, true)
+	args, err := internal.BuildHExpireArgs(key, expireTime, fields, opts, true)
 	if err != nil {
 		return nil, err
 	}
@@ -2079,11 +2076,11 @@ func (client *baseClient) HPExpire(
 //
 // Return value:
 //
-//	An array of integers indicating the result for each field:
-//		-2: Field does not exist in the hash, or hash is empty.
-//		0: The specified condition was not met.
-//		1: The expiration time was applied.
-//		2: When called with 0 milliseconds or past Unix time.
+//   - An array of integers indicating the result for each field:
+//   - -2: Field does not exist in the hash, or hash is empty.
+//   - 0: The specified condition was not met.
+//   - 1: The expiration time was applied.
+//   - 2: When called with 0 milliseconds or past Unix time.
 //
 // [valkey.io]: https://valkey.io/commands/hpexpireat/
 func (client *baseClient) HPExpireAt(
@@ -2093,7 +2090,7 @@ func (client *baseClient) HPExpireAt(
 	fields []string,
 	opts options.HExpireOptions,
 ) ([]int64, error) {
-	args, err := options.BuildHExpireArgs(key, expireTime, fields, opts, true)
+	args, err := internal.BuildHExpireArgs(key, expireTime, fields, opts, true)
 	if err != nil {
 		return nil, err
 	}
@@ -2123,14 +2120,14 @@ func (client *baseClient) HPExpireAt(
 //
 // Return value:
 //
-//	An array of integers indicating the result for each field:
-//		-2: Field does not exist in the hash, or hash does not exist.
-//		-1: Field exists but has no expiration.
-//		1: The expiration was successfully removed from the field.
+//   - An array of integers indicating the result for each field:
+//   - -2: Field does not exist in the hash, or hash does not exist.
+//   - -1: Field exists but has no expiration.
+//   - 1: The expiration was successfully removed from the field.
 //
 // [valkey.io]: https://valkey.io/commands/hpersist/
 func (client *baseClient) HPersist(ctx context.Context, key string, fields []string) ([]int64, error) {
-	args, err := options.BuildHPersistArgs(key, fields)
+	args, err := internal.BuildHPersistArgs(key, fields)
 	if err != nil {
 		return nil, err
 	}
@@ -2159,12 +2156,14 @@ func (client *baseClient) HPersist(ctx context.Context, key string, fields []str
 //
 // Return value:
 //
-//	An array of integers indicating the TTL for each field in seconds:
-//	Positive number: remaining TTL, -1: field exists but has no expiration, -2: field doesn't exist.
+//   - An array of integers indicating the TTL for each field in seconds:
+//   - Positive number: remaining TTL.
+//   - -1: field exists but has no expiration.
+//   - -2: field doesn't exist.
 //
 // [valkey.io]: https://valkey.io/commands/httl/
 func (client *baseClient) HTtl(ctx context.Context, key string, fields []string) ([]int64, error) {
-	args, err := options.BuildHTTLAndExpireTimeArgs(key, fields)
+	args, err := internal.BuildHTTLAndExpireTimeArgs(key, fields)
 	if err != nil {
 		return nil, err
 	}
@@ -2193,12 +2192,14 @@ func (client *baseClient) HTtl(ctx context.Context, key string, fields []string)
 //
 // Return value:
 //
-//	An array of integers indicating the TTL for each field in milliseconds:
-//	Positive number: remaining TTL, -1: field exists but has no expiration, -2: field doesn't exist.
+//   - An array of integers indicating the TTL for each field in milliseconds:
+//   - Positive number: remaining TTL.
+//   - -1: field exists but has no expiration.
+//   - -2: field doesn't exist.
 //
 // [valkey.io]: https://valkey.io/commands/hpttl/
 func (client *baseClient) HPTtl(ctx context.Context, key string, fields []string) ([]int64, error) {
-	args, err := options.BuildHTTLAndExpireTimeArgs(key, fields)
+	args, err := internal.BuildHTTLAndExpireTimeArgs(key, fields)
 	if err != nil {
 		return nil, err
 	}
@@ -2227,12 +2228,14 @@ func (client *baseClient) HPTtl(ctx context.Context, key string, fields []string
 //
 // Return value:
 //
-//	An array of integers indicating the expiration timestamp for each field in seconds:
-//	Positive number: expiration timestamp, -1: field exists but has no expiration, -2: field doesn't exist.
+//   - An array of integers indicating the expiration timestamp for each field in seconds:
+//   - Positive number: expiration timestamp.
+//   - -1: field exists but has no expiration.
+//   - -2: field doesn't exist.
 //
 // [valkey.io]: https://valkey.io/commands/hexpiretime/
 func (client *baseClient) HExpireTime(ctx context.Context, key string, fields []string) ([]int64, error) {
-	args, err := options.BuildHTTLAndExpireTimeArgs(key, fields)
+	args, err := internal.BuildHTTLAndExpireTimeArgs(key, fields)
 	if err != nil {
 		return nil, err
 	}
@@ -2262,12 +2265,14 @@ func (client *baseClient) HExpireTime(ctx context.Context, key string, fields []
 //
 // Return value:
 //
-//	An array of integers indicating the expiration timestamp for each field in milliseconds:
-//	Positive number: expiration timestamp, -1: field exists but has no expiration, -2: field doesn't exist.
+//   - An array of integers indicating the expiration timestamp for each field in milliseconds:
+//   - Positive number: expiration timestamp.
+//   - -1: field exists but has no expiration.
+//   - -2: field doesn't exist.
 //
 // [valkey.io]: https://valkey.io/commands/hpexpiretime/
 func (client *baseClient) HPExpireTime(ctx context.Context, key string, fields []string) ([]int64, error) {
-	args, err := options.BuildHTTLAndExpireTimeArgs(key, fields)
+	args, err := internal.BuildHTTLAndExpireTimeArgs(key, fields)
 	if err != nil {
 		return nil, err
 	}
