+++ go/options/hash_expiration_options.go
@@ -5,9 +5,6 @@
 package options
 
 import (
-	"errors"
-	"time"
-
 	"github.com/valkey-io/valkey-glide/go/v2/constants"
 	"github.com/valkey-io/valkey-glide/go/v2/internal/utils"
 )
@@ -43,12 +40,14 @@ func (opts HSetExOptions) SetExpiry(expiry *Expiry) HSetExOptions {
 func (opts *HSetExOptions) ToArgs() ([]string, error) {
 	args := []string{}
 
-	// Add conditional set options
-	conditionStr, err := opts.ConditionalSet.ToString()
-	if err != nil {
-		return nil, err
+	// Add conditional set options only if set
+	if opts.ConditionalSet != "" {
+		conditionStr, err := opts.ConditionalSet.ToString()
+		if err != nil {
+			return nil, err
+		}
+		args = append(args, conditionStr)
 	}
-	args = append(args, conditionStr)
 
 	// Add expiry options
 	if opts.Expiry != nil {
@@ -136,139 +135,14 @@ func (opts HExpireOptions) SetExpireCondition(condition constants.ExpireConditio
 func (opts *HExpireOptions) ToArgs() ([]string, error) {
 	args := []string{}
 
-	// Add expire condition options
-	conditionStr, err := opts.ExpireCondition.ToString()
-	if err != nil {
-		return nil, err
-	}
-	args = append(args, conditionStr)
-
-	return args, nil
-}
-
-// Helper functions for building command arguments
-
-// buildFieldsArgs builds the FIELDS portion of hash field expiration commands.
-func buildFieldsArgs(fields []string) []string {
-	if len(fields) == 0 {
-		return []string{}
-	}
-
-	args := []string{"FIELDS", utils.IntToString(int64(len(fields)))}
-	args = append(args, fields...)
-	return args
-}
-
-// BuildHSetExArgs builds arguments for HSETEX command.
-func BuildHSetExArgs(key string, fieldsAndValues map[string]string, options HSetExOptions) ([]string, error) {
-	if len(fieldsAndValues) == 0 {
-		return nil, errors.New("fieldsAndValues map cannot be empty")
-	}
-
-	args := []string{key}
-
-	optionArgs, err := options.ToArgs()
-	if err != nil {
-		return nil, err
-	}
-	args = append(args, optionArgs...)
-
-	// Add FIELDS keyword and count
-	args = append(args, "FIELDS", utils.IntToString(int64(len(fieldsAndValues))))
-
-	// Add field-value pairs
-	for field, value := range fieldsAndValues {
-		args = append(args, field, value)
-	}
-
-	return args, nil
-}
-
-// BuildHGetExArgs builds arguments for HGETEX command.
-func BuildHGetExArgs(key string, fields []string, options HGetExOptions) ([]string, error) {
-	if len(fields) == 0 {
-		return nil, errors.New("fields array cannot be empty")
-	}
-
-	args := []string{key}
-
-	optionArgs, err := options.ToArgs()
-	if err != nil {
-		return nil, err
-	}
-	args = append(args, optionArgs...)
-
-	// Add FIELDS portion
-	args = append(args, buildFieldsArgs(fields)...)
-
-	return args, nil
-}
-
-// BuildHExpireArgs builds arguments for HEXPIRE/HEXPIREAT/HPEXPIRE/HPEXPIREAT commands.
-// Accepts either time.Duration (for relative expiration) or time.Time (for absolute expiration).
-// The useMilliseconds flag determines whether to convert to milliseconds (true) or seconds (false).
-func BuildHExpireArgs(
-	key string,
-	expireTime interface{},
-	fields []string,
-	options HExpireOptions,
-	useMilliseconds bool,
-) ([]string, error) {
-	if len(fields) == 0 {
-		return nil, errors.New("fields array cannot be empty")
-	}
-
-	var timeValue int64
-	switch t := expireTime.(type) {
-	case time.Duration:
-		if useMilliseconds {
-			timeValue = t.Milliseconds()
-		} else {
-			timeValue = int64(t.Seconds())
+	// Add expire condition options only if set
+	if opts.ExpireCondition != "" {
+		conditionStr, err := opts.ExpireCondition.ToString()
+		if err != nil {
+			return nil, err
 		}
-	case time.Time:
-		if useMilliseconds {
-			timeValue = t.UnixMilli()
-		} else {
-			timeValue = t.Unix()
-		}
-	default:
-		return nil, errors.New("expireTime must be either time.Duration or time.Time")
+		args = append(args, conditionStr)
 	}
 
-	args := []string{key, utils.IntToString(timeValue)}
-
-	optionArgs, err := options.ToArgs()
-	if err != nil {
-		return nil, err
-	}
-	args = append(args, optionArgs...)
-
-	// Add FIELDS portion
-	args = append(args, buildFieldsArgs(fields)...)
-
-	return args, nil
-}
-
-// BuildHPersistArgs builds arguments for HPERSIST command.
-func BuildHPersistArgs(key string, fields []string) ([]string, error) {
-	if len(fields) == 0 {
-		return nil, errors.New("fields array cannot be empty")
-	}
-
-	args := []string{key}
-	args = append(args, buildFieldsArgs(fields)...)
-	return args, nil
-}
-
-// BuildHTTLAndExpireTimeArgs builds arguments for hash field TTL and expiration time query commands.
-// Supports HTTL, HPTTL, HEXPIRETIME, and HPEXPIRETIME commands that check existing time information.
-func BuildHTTLAndExpireTimeArgs(key string, fields []string) ([]string, error) {
-	if len(fields) == 0 {
-		return nil, errors.New("fields array cannot be empty")
-	}
-
-	args := []string{key}
-	args = append(args, buildFieldsArgs(fields)...)
 	return args, nil
 }
